+++ docs/changelog.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [v0.0.1-beta.1] - 04.07.2023
 
 ### Added
@@ -90,3 +92,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Export InputStatus
 - Change style of skeleton of SCn
+
+## [v0.1.1] - 13.12.2023
+
+### Fixed
+
+- Passing font weight and color to placeholder
+
+### Added
+
+- Playground for select
+- Toasts Animation
+- Custom Toast positions
+- Tooltip classname